+++ server_modules/mongodb.js
@@ -1,0 +1,7 @@
+/*引入*/
+var mongoose = require('mongoose')
+// 用于异步回调
+mongoose.Promise = require('bluebird')
+global.db = mongoose.connect("mongodb://localhost:27017/vuechat")
+
+module.exports = mongoose;
